Updated capsensor and radio current draws for version 122
I simply increased the radio draw to 100uA on the radio, and
increased the capsensor static draw to 350uA.  This is a bit
high for each of these values, but it isn't unrealistic.  And
the verification current now matches what Frank gets.

Unfortunately, we measure closer to 200uA static current.  We may
have some problems with our cheap DMM relative to Frank's
sourcemeter.